fix: "too many tick handlers" crash in BlastEm with non-MW ROM
This disables MegaWiFi functionality entirely in non-MW ROM. Turns out the MW detection logic doesn't work properly in BlastEm when the header is SEGA_SSF. The MW hw is always detected and seems to trigger stack corruption calling `mw_udp_set`. Disabling all of it entirely if SEGA_MEGAWIFI is not set makes sense to avoid this fuss.